better change to oop
